feat(E12-S02a): customer-app Hindi sweep residuals — Noto Devanagari font + Typography fallback + Detekt-equivalent gate
- Add NotoSansDevanagari-Regular.ttf (647 KB, OFL-1.1) to design-system/src/main/res/font/
- Move OFL license to design-system/LICENSES/ (Android res/font/ only accepts .ttf/.xml)
- Wire NotoSansDevanagariFontFamily + HomeservicesDevanagariFontFamily (Geist + Noto per-glyph fallback)
  using FontLoadingStrategy.OptionalLocal on all Noto entries so Paparazzi JVM tests don't crash
- Apply HomeservicesDevanagariFontFamily to all HomeservicesTypography text styles (headline, title, body, label, display)
- Add verifyNoEnglishTextLiterals Gradle task wired into check; 0 violations found post-sweep
- Fix PriceApprovalScreen.kt: sumOf { it.price } returns Int but formatInr expects Long — add .toLong()
- Fix LocaleRepositoryImplTest: update default-locale assertion from "en" to "hi" (ADR-0018 Ayodhya default)
- ktlint auto-format: spacing-between-declarations-with-comments in LocaleRepositoryImpl.kt,
  import-ordering in PriceApprovalScreen.kt, string-template in CurrencyFormat.kt
- pre-codex-smoke.sh exits 0 (344 tests pass, kover >= 80%, ktlint clean, assembleDebug clean)

+++ customer-app/app/src/main/kotlin/com/homeservices/customer/ui/booking/PriceApprovalScreen.kt
@@ -30,8 +30,8 @@ import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.homeservices.customer.R
 import com.homeservices.customer.domain.booking.model.AddOnDecision
-import com.homeservices.customer.ui.util.formatInr
 import com.homeservices.customer.domain.booking.model.PendingAddOn
+import com.homeservices.customer.ui.util.formatInr
 import com.homeservices.designsystem.components.HsPriceText
 import com.homeservices.designsystem.components.HsPrimaryButton
 import com.homeservices.designsystem.components.HsSecondaryButton
@@ -171,7 +171,7 @@ private fun TotalDecisionSummary(
     addOns: List<PendingAddOn>,
     decisions: Map<String, Boolean>,
 ) {
-    val approvedTotal = addOns.filter { decisions[it.name] == true }.sumOf { it.price }
+    val approvedTotal = addOns.filter { decisions[it.name] == true }.sumOf { it.price.toLong() }
     Surface(
         modifier = Modifier.fillMaxWidth(),
         shape = MaterialTheme.shapes.medium,
